requirements.txt
# Web framework
streamlit>=1.30.0

# YouTube transcript fetching
youtube-transcript-api>=0.6.0

# ASR fallback (Apple Silicon only)
mlx-whisper>=0.3.0

# Claude API for summarization
anthropic>=0.18.0

# Video/audio processing
yt-dlp>=2024.0.0

# Environment variables
python-dotenv>=1.0.0

# Database ORM
sqlalchemy>=2.0.0
